config: Wrap config library in an interface
